+++ workspace/node/bookmarks/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./bookmark";
+export * from "./builders";
 export * from "./parsers";